Include all Renpy base files
This allows mods to be installed easily on Ren'Py 7 (and 8 if applied to the Python-3 branch) by only needing to copy the 3 DDLC rpa files into the /game/ folder. This still follows the Team Salvato IP Guidelines, as it would only make sure that all of the open source Ren'Py files are added to the project.

Previously, you would have to move the entire mod into the DDLC folder, as it merges the new Ren'Py and Python files with the old ones. This is only necessary because DDLC's Ren'Py files include default assets such as fonts that aren't included in current mod builds.

Adding the 'all' function to packages allows these to be included, which is required for at least one specific new feature in Ren'Py 8.2/7.7, being the native emoji support (the emojis use a font in Ren'Py files).

Because of this change, files that Android needs have been changed from 'all' to 'android', to make sure that they aren't included in newer builds.